Mechaduino: PID fixes
Setting the stepper position during PID init is a bad idea, as it results in an initial phase shift when entering the PID loop.  Now an offset is calculated and applied to the encoder phase.

Its concerning that the loop was unable to self-correct after the initial move, which suggests I still have a logic error in there.