servo_stepper:  attempt to fix stats reporting
Currently when the error goes negative it reports a very large negative number back to Klipper. I think that it is likely due to the protocol and this patch
will work around it.

This patch also re-enables the attempt at holding in the PID loop when the current error is below a threshold of +-16 microsteps of the current error.  I may
remove this as an unnecessary later, particularly after the Kalman Filter (or Alpha Beta) is implemented.