servo_stepper: track encoder difference
This is similar to a previous implementation where the encoder position is tracked and the phase difference is computed from the encoder difference.

This implementation does two things that the previous did not do:  First it computes the current phase from the current position and issues a move relative to that.  Second it offsets the encoder position based on what should be phase 0, this is likely what caused issues previously.