servo_stepper:  Attempts to fix issues when intializing PID
There seems to be some issue when enabling the stepper, sometimes it loses its mind, sometimes its ok.  It seems to affect the stepper the most after its been moved when disabled, which suggests that I needed to delay init a bit to make sure I get the correct phase.

Also have converted the init procedure to use a moving average.